// src/app/api/employee-client/model/employeeRequestDto.ts
/**
 * Employee Service API
 *
 * Contact: support@employee-service.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AddressDto } from './addressDto';
import { EmergencyContactDto } from './emergencyContactDto';
import { ContactInformationDto } from './contactInformationDto';


/**
 * Request object for updating and creating an employee
 */
export interface EmployeeRequestDto { 
    /**
     * Employee\'s first name
     */
    firstName: string;
    /**
     * Employee\'s middle name
     */
    middleName?: string | null;
    /**
     * Employee\'s last name
     */
    lastName: string;
    /**
     * Employee\'s job title
     */
    jobTitle: string;
    /**
     * URL of employee\'s profile image
     */
    imageUrl?: string | null;
    /**
     * Date when employee was hired
     */
    hiredDate: string;
    /**
     * Employee\'s birth date
     */
    birthDate: string;
    addressDto: AddressDto;
    contactInformationDto: ContactInformationDto;
    emergencyContactDto: EmergencyContactDto;
}